feat: Add Generic Status Type master with full CRUD operations
Backend:
- Created GenericStatusType entity with JPA annotations
- Implemented DTOs (Request/Response) with validation
- Created repository, service, and controller layers
- Added pagination, search, and list endpoints
- Created Postman collection for API testing

Frontend:
- Created complete feature module with API integration
- Implemented data table with sorting and pagination
- Added create/edit drawer with form validation
- Implemented delete confirmation dialog
- Integrated with React Query for state management
- Added route and navigation in sidebar

UI Improvements:
- Reorganized sidebar with 3-level nested categories
- Added Master Data parent menu with subcategories:
  - Location, Financial, Vendor, Asset, Activity, System
- Added icons to all menu items
- Increased sidebar width from 16rem to 20rem for better readability

Documentation:
- Updated api-design.md with Generic Status Type completion
- Fixed Postman collections for Activity and Activities APIs

+++ server-api/epsoneapi/src/main/java/com/eps/module/api/epsone/activity/service/ActivityServiceImpl.java
@@ -103,7 +103,7 @@ public class ActivityServiceImpl implements ActivityService {
                 .orElseThrow(() -> new IllegalArgumentException("Activity not found with ID: " + id));
 
         // Check if activity is being used by Activities records
-        List<com.eps.module.activity.Activities> dependentActivities = activitiesRepository.findByActivityId(id.intValue());
+        List<com.eps.module.activity.Activities> dependentActivities = activitiesRepository.findByActivityId((long) id.intValue());
         if (!dependentActivities.isEmpty()) {
             String activityNames = dependentActivities.stream()
                 .limit(3) // Show max 3 names
